+++ examples/v2016.rs
@@ -2,7 +2,7 @@
 // Copyright (c) 2023 Nathan Fiedler
 //
 use clap::{arg, command, value_parser, Arg};
-use fastcdc::chunker::v2016::*;
+use fastcdc::v2016::*;
 use memmap::MmapOptions;
 use std::fs::File;